requirements.txt
Django==4.2.11
djangorestframework==3.14.0
Pillow==10.3.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
stripe==7.0.0
python-dotenv==1.0.0
